add error handling
