+++ back-end/api/routes/admin.js
@@ -1,0 +1,8 @@
+const express = require('express');
+
+const adminController = require('../controllers/admin');
+
+const router = express.Router();
+
+
+module.exports = router;
