+++ back-end/api/server.js
@@ -1,0 +1,8 @@
+const http = require("http");
+const app = require("./app");
+
+const port = process.env.PORT || 4000;
+
+const server = http.createServer(app);
+
+server.listen(port);
